chore(front): move FileUploadComponent to components/
