+++ apps/frontend/src/app/components/components.module.ts
@@ -20,6 +20,7 @@ import { UserSearchComponent } from './user-search/user-search.component';
 import { UserSearchResultComponent } from './user-search/user-search-result.component';
 import { NbFormFieldModule } from '@nebular/theme';
 import { NgxPaginationModule } from 'ngx-pagination';
+import { FileUploadComponent } from './file-upload/file-upload.component';
 
 const COMPONENTS = [
   HeaderComponent,
@@ -33,7 +34,8 @@ const COMPONENTS = [
   ConfirmDialogComponent,
   DeleteUserDialogComponent,
   UserSearchComponent,
-  UserSearchResultComponent
+  UserSearchResultComponent,
+  FileUploadComponent
 ];
 
 @NgModule({
