+++ apps/frontend/src/app/components/file-upload/file-upload.component.scss
